requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
djangorestframework==3.14.0
requests==2.31.0